modified double to bigdecimal

+++ src/main/java/ru/clevertec/hardziyevich/api/certificate/GiftCertificatePostDto.java
@@ -2,6 +2,8 @@ package ru.clevertec.hardziyevich.api.certificate;
 
 import lombok.*;
 import ru.clevertec.hardziyevich.api.tag.TagPostDto;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,8 +12,7 @@ public class GiftCertificatePostDto {
 
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private Integer duration;
     private List<TagPostDto> tags;
-
 }
